Moving lock status logic out of CreatorLock
